DRY out the 2 runs pages
