+++ ui/apps/dashboard/src/components/Runs/index.ts
@@ -1,0 +1,1 @@
+export { Runs } from './Runs';
