+++ ui/apps/dashboard/src/gql/gql.ts
@@ -101,11 +101,9 @@ const documents = {
     "\n  query GetRunTrace($envID: ID!, $runID: String!) {\n    workspace(id: $envID) {\n      run(runID: $runID) {\n        function {\n          app {\n            name\n            externalID\n          }\n          id\n          name\n          slug\n        }\n        trace {\n          ...TraceDetails\n          childrenSpans {\n            ...TraceDetails\n            childrenSpans {\n              ...TraceDetails\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetRunTraceDocument,
     "\n  query TraceResult($envID: ID!, $traceID: String!) {\n    workspace(id: $envID) {\n      runTraceSpanOutputByID(outputID: $traceID) {\n        data\n        error {\n          message\n          name\n          stack\n        }\n      }\n    }\n  }\n": types.TraceResultDocument,
     "\n  query GetRunTraceTrigger($envID: ID!, $runID: String!) {\n    workspace(id: $envID) {\n      runTrigger(runID: $runID) {\n        IDs\n        payloads\n        timestamp\n        eventName\n        isBatch\n        batchID\n        cron\n      }\n    }\n  }\n": types.GetRunTraceTriggerDocument,
-    "\n  query GetEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n": types.GetEnvRunsDocument,
-    "\n  query CountEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n": types.CountEnvRunsDocument,
+    "\n  query GetRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n": types.GetRunsDocument,
+    "\n  query CountRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n": types.CountRunsDocument,
     "\n  query AppFilter($envSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n      }\n    }\n  }\n": types.AppFilterDocument,
-    "\n  query GetFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n": types.GetFnRunsDocument,
-    "\n  query CountFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n": types.CountFnRunsDocument,
     "\n  query GetDeployss($environmentID: ID!) {\n    deploys(workspaceID: $environmentID) {\n      id\n      appName\n      authorID\n      checksum\n      createdAt\n      error\n      framework\n      metadata\n      sdkLanguage\n      sdkVersion\n      status\n\n      deployedFunctions {\n        id\n        name\n      }\n\n      removedFunctions {\n        id\n        name\n      }\n    }\n  }\n": types.GetDeployssDocument,
     "\n  query GetEnvironments {\n    workspaces {\n      id\n      name\n      slug\n      parentID\n      test\n      type\n      webhookSigningKey\n      createdAt\n      isArchived\n      isAutoArchiveEnabled\n      lastDeployedAt\n    }\n  }\n": types.GetEnvironmentsDocument,
     "\n  query GetEnvironmentBySlug($slug: String!) {\n    envBySlug(slug: $slug) {\n      id\n      name\n      slug\n      parentID\n      test\n      type\n      createdAt\n      lastDeployedAt\n      isArchived\n      isAutoArchiveEnabled\n      webhookSigningKey\n    }\n  }\n": types.GetEnvironmentBySlugDocument,
@@ -490,23 +488,15 @@ export function graphql(source: "\n  query GetRunTraceTrigger($envID: ID!, $runI
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query GetEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query GetRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query CountEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"): (typeof documents)["\n  query CountEnvRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query CountRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"): (typeof documents)["\n  query CountRuns(\n    $appIDs: [UUID!]\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          appIDs: $appIDs\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query AppFilter($envSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n      }\n    }\n  }\n"): (typeof documents)["\n  query AppFilter($envSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      apps {\n        externalID\n        id\n        name\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n    $functionRunCursor: String = null\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n        after: $functionRunCursor\n      ) {\n        edges {\n          node {\n            app {\n              externalID\n              name\n            }\n            cronSchedule\n            eventName\n            function {\n              name\n              slug\n            }\n            id\n            isBatch\n            queuedAt\n            endedAt\n            startedAt\n            status\n          }\n        }\n        pageInfo {\n          hasNextPage\n          hasPreviousPage\n          startCursor\n          endCursor\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query CountFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"): (typeof documents)["\n  query CountFnRuns(\n    $environmentID: ID!\n    $startTime: Time!\n    $endTime: Time\n    $status: [FunctionRunStatus!]\n    $timeField: RunsOrderByField!\n    $functionSlug: String!\n  ) {\n    environment: workspace(id: $environmentID) {\n      runs(\n        filter: {\n          from: $startTime\n          until: $endTime\n          status: $status\n          timeField: $timeField\n          fnSlug: $functionSlug\n        }\n        orderBy: [{ field: $timeField, direction: DESC }]\n      ) {\n        totalCount\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
